fix: cast CPUAccessFlags to u32 for Windows build

+++ src/capture/windows.rs
@@ -129,7 +129,7 @@ impl DxgiCapture {
             },
             Usage: D3D11_USAGE_STAGING,
             BindFlags: 0,
-            CPUAccessFlags: D3D11_CPU_ACCESS_READ.0,
+            CPUAccessFlags: D3D11_CPU_ACCESS_READ.0 as u32,
             MiscFlags: 0,
         };
